split etl based on entities: application, noi, documents
