+++ bin/migrate-oats-data/applications/__init__.py
@@ -1,2 +1,8 @@
 from .base_applications import *
 from .app_prep import process_alcs_application_prep_fields
+from .submissions import process_alcs_app_submissions
+from .application_submission_status_email import (
+    process_application_submission_status_emails,
+    clean_application_submission_status_emails,
+)
+from .base_applications import process_applications
