+++ bin/migrate-oats-data/menu/__init__.py
@@ -1,0 +1,2 @@
+from .menu import setup_menu_args_parser
+from .commands import *
