+++ bin/migrate-oats-data/menu/command_parser/__init__.py
@@ -1,0 +1,3 @@
+from .application_command_parser import *
+from .notice_of_intent_command_parser import *
+from .document_command_parser import *
