+++ bin/migrate-oats-data/menu/commands/__init__.py
@@ -1,0 +1,5 @@
+from .import_all import import_all
+from .clean_all import clean_all
+from .documents import *
+from .applications import *
+from .notice_of_intents import *
