+++ bin/migrate-oats-data/noi/__init__.py
@@ -1,1 +1,2 @@
-from .notice_of_intent_init import *
+from .notice_of_intent_init import init_notice_of_intents, clean_notice_of_intents
+from .notice_of_intent_migration import process_notice_of_intent
