+++ bin/migrate-oats-data/noi/oats_to_alcs_notice_of_intent_table_etl/__init__.py
@@ -1,1 +1,3 @@
-from .oats_to_alcs_notice_of_intent_table_etl import process_alcs_notice_of_intent_fee_fields
+from .oats_to_alcs_notice_of_intent_table_etl import (
+    process_alcs_notice_of_intent_base_fields,
+)
